baboteek-api/pyproject.toml
[project]
name = "baboteek-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "baboteek-core",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.136.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[tool.uv.sources]
baboteek-core = { workspace = true }
